backend/requirements.txt
absl-py==2.1.0
accelerate==1.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
altgraph @ file:///AppleInternal/Library/BuildRoots/860631e9-c1c5-11ee-98ee-b6ef2fd8d87b/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
async-timeout==5.0.1
attrs==24.3.0
av==15.1.0
beautifulsoup4==4.12.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.0
click==8.1.8
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.0
ctranslate2==4.7.1
cycler==0.12.1
datasets==3.2.0
debugpy==1.8.11
decorator==5.1.1
dill==0.3.8
diskcache==5.6.3
dnspython==2.7.0
docker-pycreds==0.4.0
eval_type_backport==0.2.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.14
faster-whisper==1.2.1
filelock==3.16.1
flatbuffers==25.12.19
fonttools==4.55.8
frozenlist==1.5.0
fsspec==2024.9.0
future @ file:///AppleInternal/Library/BuildRoots/860631e9-c1c5-11ee-98ee-b6ef2fd8d87b/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
gitdb==4.0.11
GitPython==3.1.43
google==3.0.0
grpcio==1.68.1
h11==0.14.0
huggingface-hub==0.27.0
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.2
Jinja2==3.1.4
jupyter_client==8.6.3
jupyter_core==5.7.2
kagglehub==0.3.5
kiwisolver==1.4.7
llama_cpp_python==0.3.5
llvmlite==0.43.0
lxml==5.3.0
macholib @ file:///AppleInternal/Library/BuildRoots/860631e9-c1c5-11ee-98ee-b6ef2fd8d87b/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.3
numba==0.60.0
numpy==2.0.2
onnxruntime==1.19.2
openai-whisper==20250625
ordereddict==1.1
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==5.29.2
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
pycparser==2.23
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
pymongo==4.11.1
pyparsing==3.2.1
PySocks==1.7.1
pysolr==3.5.0
python-dateutil==2.9.0.post0
python-dotenv==0.5.1
pytrends==4.9.2
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.11.6
requests==2.32.5
safetensors==0.4.5
selenium==4.36.0
sentencepiece==0.2.0
sentry-sdk==2.19.2
setproctitle==1.3.4
six==1.10.0
slackclient==1.0.1
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.5
soupsieve==2.6
sseclient==0.0.27
stack-data==0.6.3
starlette==0.46.2
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tiktoken==0.12.0
tokenizers==0.21.0
torch==2.5.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.1
trio==0.31.0
trio-websocket==0.12.2
typing_extensions==4.15.0
tzdata==2024.2
urllib3==2.6.3
uvicorn==0.39.0
wandb==0.19.1
watson-developer-cloud==0.18.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
